consistent file structure
