+++ src/probnum/quad/solvers/initial_designs/_latin_design.py
@@ -3,9 +3,11 @@
 import numpy as np
 
 from probnum.quad.integration_measures import IntegrationMeasure
+
 from ._initial_design import InitialDesign
 
 
+# pylint: disable=too-few-public-methods
 class LatinDesign(InitialDesign):
     """Initial design for Bayesian quadrature that samples from a latin hypercube.
 
@@ -15,9 +17,13 @@ class LatinDesign(InitialDesign):
         The integration measure.
      rng
         A random number generator.
-   """
+    """
 
-    def __init__(self, measure: IntegrationMeasure, rng: np.random.Generator = np.random.default_rng()) -> None:
+    def __init__(
+        self,
+        measure: IntegrationMeasure,
+        rng: np.random.Generator = np.random.default_rng(),
+    ) -> None:
         # Todo: measure domain needs to be cube
         super().__init__(measure=measure)
         self.rng = rng
